refactor(tabs): split useTabs into tabOps/ submodules
The 997-LOC src/hooks/useTabs.ts was the frontend's biggest state knot —
agent/shell/editor tab lifecycle, root-state mirror writes, terminal
replay dispatch, auto-restore, close+reopen, Monaco buffer cleanup, and
bridge invokes were all in one closure. Split into one facade +
nine focused submodules, same pattern as the existing projectOps/
directory.

Layout:

  src/hooks/useTabs.ts                — facade (147 LOC): ref creation,
                                        action wiring, public re-exports
  src/hooks/tabOps/types.ts           — UseTabsContext, UseTabsActions,
                                        DiscoveredSession, NotificationInput
  src/hooks/tabOps/constants.ts       — TAB_MIRROR_KEYS, TERMINAL_REPLAY_MAX,
                                        CLOSED_TAB_STACK_MAX, VALID_SHARE_MODES
  src/hooks/tabOps/helpers.ts         — pure data helpers (session labels,
                                        model/cwd resolution, recentSessions
                                        projection, editorLabelForPath)
  src/hooks/tabOps/mutations.ts       — useMutations + dispatchTerminalReplay
                                        (updateTab, updateActiveTab,
                                        applyShareModeToTab, setActiveTab,
                                        setActiveSubTab)
  src/hooks/tabOps/agentTab.ts        — useNewTab (agent-tab creation with
                                        bridge tab_open + scroll-to-match seed)
  src/hooks/tabOps/shellTab.ts        — useNewShellTab (panel-only, seeds
                                        share mode atomically inside shell_open)
  src/hooks/tabOps/editorTab.ts       — useEditorTabActions (newEditorTab,
                                        updateEditorMeta, toggleEditorPreview,
                                        renameEditorTabsForPath)
  src/hooks/tabOps/closeTab.ts        — useCloseTabActions (closeTab,
                                        closeTabNow, pushClosedTab,
                                        reopenLastClosedTab, plus
                                        closeEditorTabsForPath which lives
                                        here so the bulk-close honors the
                                        dirty-buffer confirm prompt)
  src/hooks/tabOps/autoRestore.ts     — useAutoRestoreDiscoveredSessions

The facade re-exports the public surface — TAB_MIRROR_KEYS,
TERMINAL_REPLAY_MAX, cwdForNewTab, modelForNewProjectTab,
recentSessionItemFromClosedTab, UseTabsContext, UseTabsActions, and
useTabs itself — so every caller (App.tsx, bridgeMessageHandlers/*,
useOsEdges, useSessionPersistence, projectOps/tabBuckets,
useTabs.test) keeps the same import path. No external file changes.

Factory functions are named `useX` to match the existing projectOps/
convention (useProjectStore, useWorktreeOperations); React's lint rule
trusts the hook-naming convention when refs are passed in.

Public-surface guarantee verified by leaving useTabs.test.ts untouched
— it imports cwdForNewTab, modelForNewProjectTab, and
recentSessionItemFromClosedTab from "./useTabs", all three of which the
facade re-exports from tabOps/helpers.

bunx tsc --noEmit                 clean
bunx eslint . --max-warnings=5    5 pre-existing warnings (none from this PR)
bunx vitest run                   148/148 files, 1124/1124 tests
cargo test --lib                  158/158 (untouched)
